[Versioning]

+++ include/erc/package.h
@@ -7,10 +7,10 @@ namespace erc {
   struct package
   {
     const std::string name;
-    const size_t size;
+    const uint size;
     const erc::embedded_file * const * const embedded_files;
 
-    inline package( const std::string & _name, const size_t _size, const erc::embedded_file * const * const _embedded_files ) :
+    inline package( const std::string & _name, const uint _size, const erc::embedded_file * const * const _embedded_files ) :
       name( std::move( _name ) ), size( std::move( _size ) ), embedded_files( std::move( _embedded_files ) )
     {}
   };
